Created Extensions to lighten startup.cs
