+++ ArchaicQuestII.API/Configuration/IoC/GameLogicExtensions/CombatExtensions.cs
@@ -1,0 +1,16 @@
+﻿using ArchaicQuestII.GameLogic.Combat;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace ArchaicQuestII.API.Configuration.IoC.GameLogicExtensions
+{
+    public static class CombatExtensions
+    {
+        public static IServiceCollection AddCombatLogic(this IServiceCollection services)
+        {
+            services.AddSingleton<ICombat, Combat>();
+            services.AddSingleton<IFormulas, Formulas>();
+
+            return services;
+        }
+    }
+}
