+++ ArchaicQuestII.API/Helpers/ExtensionMethods.cs
@@ -1,6 +1,11 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using ArchaicQuestII.API.Entities;
+using ArchaicQuestII.GameLogic.Core;
+using ArchaicQuestII.GameLogic.Hubs.Telnet;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace ArchaicQuestII.API.Helpers
 {
@@ -17,6 +22,21 @@ namespace ArchaicQuestII.API.Helpers
 
             user.Password = null;
             return user;
+        }
+
+        public static void StartLoops(this IApplicationBuilder app)
+        {
+            var loop = app.ApplicationServices.GetRequiredService<IGameLoop>();
+
+            Task.Run(TelnetHub.Instance.ProcessConnections);
+            Task.Run(loop.UpdateTime);
+            Task.Run(loop.UpdateCombat);
+            Task.Run(loop.UpdatePlayers);
+            Task.Run(loop.UpdatePlayerLag);
+            Task.Run(loop.UpdateRoomEmote).ConfigureAwait(false);
+            Task.Run(loop.UpdateMobEmote).ConfigureAwait(false);
+            Task.Run(loop.UpdateWorldTime).ConfigureAwait(false);
+            Task.Run(loop.Tick);
         }
     }
 }
